pyproject.toml
[tool.poetry]
name = "bi-retail"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "^3.5.3"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
faker = "^33.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
